Dropped delimited period completely from FishingTripEntity
The fishing trip table previously contained three sets of dates: occurrence, delimited period and calculated start and end dates. The calculated start and end dates are (as the name suggests) calculated from occurrence and the delimited period. It is therefor not necessary to keep all of them and I opted for simplifying the data model by removing the delimited period. Reason is we have start and end in the calculated fields and the duration which is part of delimited period can easily be calculated and there is no need to query for it.

Secondly, while doing the unit tests for this I discovered I had an issue with time zones. This is due to us doing most of our testing with arquillian which runs on a wildfly that has the global timezone set to UTC. To deal with this I decided to set the timezone in the setup for each unit test to "simulate" the environment we have in wildfly.