Added methods for getting start and end dates from delimited period.
Added methods that calculates the start and end date of a delimited period if the actual value isn't present. For example, if start date is not present but duration and end date is, it will calculate the start date and return that value. This does not do any specific handling for the situation where both start and end date are set but are equal, it will simply return both those values.