Removed delimited period from fishing activity.
Because of how the new conversion works I had to ignore delimited periods when doing xml comparisons between the original xml document and the one you get when mapping an entity back to flux. The reason is that the mapping now always tries to add a delimited period back to the flux object since it doesn't know if it was there or not.