Fixed mapping compilation problems.
There were two problems with the mappers. One was that there was no mapping for FaReportIdentifierEntity and RelatedReportDto. This was easily fixed by adding those methods to the BaseMapper class. The second was an ambiguous mapping error and this I'm not sure what the root problem was. A quick googling pointed to a problem with the target members being String and mapstruct not picking the right mapping methods. But since there were qualifiers on them it should have worked (?). The fix was instead to use the expression attribute and actually use the method we tried to point to with the qualifier.