Added changesets for dropping foreign key constraints.
We tried to modify the data type of a column that was used for a foreign key constraint and also tried to drop a table that had a foreign key reference to it. Fixed a couple of queries.
Tests are still broken due to references to flux report document in several places.