Removed manually created getLocations method
This method used to be named getLocations_ which doesn't comply with the Java naming standard. I originally thought this overwrote an existing getLocations but the lombok generated getter is not named that. It turns out that mapstruct seems to use all getters, regardless if they have been listed or not.