Updated queries for getting fishing activities to order by date and purpose code.
The current implementation of the queries returned a list that was sorted on type code which effectively meant it return it in alphabetical order. A more natural order would be by date and time so you get the list of activities in the order they happened. This was easy enough to do until you consider that some activities seem to happen in pairs with the same date and time. For the data we have these turned out to have purpose code 9 and 5 for new and correction respectively. So sorting both on timestamp and on reverse purpose code gives us a list sorted by date and time and if there are multiple with the same date and time it will always put new ones first (those with purpose code 9).